+++ src/Components/Cart/DisplayCart.js
@@ -5,7 +5,6 @@ import React from 'react'
 export default function DisplayCart(props) {
   return (
     <div>
-        <h2>Cart</h2>
         {props.cartArray.map(eachProduct => {
                 return(
                     <div key={eachProduct.itemId}>
